refactor(frontend): refactor library sub-page
