+++ frontend/src/components/home/Sidebar.tsx
@@ -7,6 +7,7 @@ import UserAvatar from "@/components/UserAvatar"
 import { useAuthState } from "@/context/auth"
 import { useLibraryDispatch, useLibraryState } from "@/context/library"
 
+// TODO
 function AddShelfDialog() {
     const libraryDispatch = useLibraryDispatch()
