+++ frontend/src/components/home/library/index.ts
@@ -1,0 +1,3 @@
+export * from "./SortPopover.tsx"
+export * from "./BookCard.tsx"
+export * from "./BooksGrid.tsx"
