+++ frontend/src/pages/home/index.ts
@@ -1,1 +1,2 @@
 export * from "./home.tsx"
+export * from "./library.tsx"
